Move CA1014 and CA1016 to their new location - these are implemented by one analyzer - deleting the stub filess for the two rules.
