+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/Core/MarkAssembliesWithClsCompliant.Fixer.cs
@@ -18,7 +18,7 @@ namespace Microsoft.ApiDesignGuidelines.Analyzers
     /// </summary>
     public abstract class MarkAssembliesWithClsCompliantFixer : CodeFixProvider
     {
-        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(MarkAssembliesWithClsCompliantAnalyzer.RuleId);
+        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(MarkAssembliesWithAttributesDiagnosticAnalyzer.CA1014RuleId);
 
         public sealed override FixAllProvider GetFixAllProvider()
         {
