+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/Core/Microsoft.ApiDesignGuidelines.Analyzers.csproj
@@ -99,6 +99,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="IdentifiersShouldNotContainTypeNames.cs" />
+    <Compile Include="MarkAssembliesWithAttributesDiagnosticAnalyzer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MicrosoftApiDesignGuidelinesAnalyzersResources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -125,9 +126,7 @@
     <Compile Include="EnumsShouldHaveZeroValue.Fixer.cs" />
     <Compile Include="AbstractTypesShouldNotHaveConstructors.cs" />
     <Compile Include="AbstractTypesShouldNotHaveConstructors.Fixer.cs" />
-    <Compile Include="MarkAssembliesWithClsCompliant.cs" />
     <Compile Include="MarkAssembliesWithClsCompliant.Fixer.cs" />
-    <Compile Include="MarkAssembliesWithAssemblyVersion.cs" />
     <Compile Include="MarkAssembliesWithAssemblyVersion.Fixer.cs" />
     <Compile Include="MarkAssembliesWithComVisible.cs" />
     <Compile Include="MarkAssembliesWithComVisible.Fixer.cs" />
