+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/MarkAssembliesWithClsCompliantTests.Fixer.cs
@@ -12,12 +12,12 @@ namespace Microsoft.ApiDesignGuidelines.Analyzers.UnitTests
     {
         protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
         {
-            return new BasicMarkAssembliesWithClsCompliantAnalyzer();
+            return new MarkAssembliesWithAttributesDiagnosticAnalyzer();
         }
 
         protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
         {
-            return new CSharpMarkAssembliesWithClsCompliantAnalyzer();
+            return new MarkAssembliesWithAttributesDiagnosticAnalyzer ();
         }
 
         protected override CodeFixProvider GetBasicCodeFixProvider()
