+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.vbproj
@@ -117,9 +117,7 @@
     <Compile Include="BasicEnumsShouldHaveZeroValue.Fixer.vb" />
     <Compile Include="BasicAbstractTypesShouldNotHaveConstructors.vb" />
     <Compile Include="BasicAbstractTypesShouldNotHaveConstructors.Fixer.vb" />
-    <Compile Include="BasicMarkAssembliesWithClsCompliant.vb" />
     <Compile Include="BasicMarkAssembliesWithClsCompliant.Fixer.vb" />
-    <Compile Include="BasicMarkAssembliesWithAssemblyVersion.vb" />
     <Compile Include="BasicMarkAssembliesWithAssemblyVersion.Fixer.vb" />
     <Compile Include="BasicDefineAccessorsForAttributeArguments.vb" />
     <Compile Include="BasicUsePropertiesWhereAppropriate.vb" />
